fixed update message

+++ server.h
@@ -24,7 +24,7 @@ struct username
 };
 
 void printUsers(struct username * users, uint16_t numberOfUsers);
-void addUserName(struct username * users, uint16_t size, char* name, int nameLen, int fd);
+void addUserName(struct username * users, uint16_t size, char* name, uint8_t nameLen, int fd);
 struct username* deleteUser(struct username * users, uint16_t size, int fd);
 
 void sendInitialHandshake(int sock);
